+++ pentair-screenlogic/CHANGELOG.md
@@ -23,3 +23,7 @@
 
 ## 1.11
 - typos and mistakes
+
+## 1.11
+- added thermostat config
+- added ability to change heater state and temperature set points